move integration tests in the folder
